refactor(naming): Renamed the base model and resource to use kebab-case instead of PascalCase.
