+++ src/resources/customers.js
@@ -1,4 +1,4 @@
-import MollieResource from 'MollieResource';
+import MollieResource from 'mollie-resource';
 import Customer from 'models/customer';
 
 /**
