+++ src/resources/issuers.js
@@ -1,4 +1,4 @@
-import MollieResource from 'MollieResource';
+import MollieResource from 'mollie-resource';
 import Issuer from 'models/issuer';
 
 /**
