+++ src/resources/methods.js
@@ -1,4 +1,4 @@
-import MollieResource from 'MollieResource';
+import MollieResource from 'mollie-resource';
 import Method from 'models/method';
 
 /**
