+++ src/resources/refunds.js
@@ -1,4 +1,4 @@
-import MollieResource from 'MollieResource';
+import MollieResource from 'mollie-resource';
 import Refund from 'models/refund';
 
 /**
